+++ kgcnn/literature/HDNNP2nd/_model.py
@@ -1,8 +1,9 @@
 from kgcnn.layers.mlp import MLP, GraphMLP, RelationalMLP
-from keras.layers import Concatenate
+from keras.layers import Concatenate, Dense, Multiply, Add
 from kgcnn.layers.pooling import PoolingNodes
 from kgcnn.layers.norm import GraphBatchNormalization
 from ._wacsf import wACSFRad, wACSFAng
+from ._layers import CorrectPartialCharges
 from ._acsf import ACSFG2, ACSFG4, ACSFConstNormalization
 
 
@@ -16,10 +17,11 @@ def model_disjoint_weighted(
         mlp_kwargs: dict = None,
         output_embedding: str = None,
         use_output_mlp: bool = None,
-        output_mlp: dict = None
+        output_mlp: dict = None,
+        predict_dipole: bool = None
 ):
     # Make input
-    node_input, xyz_input, edge_index_input, angle_index_input, batch_id_node, count_nodes = inputs
+    node_input, xyz_input, edge_index_input, angle_index_input, tot_charge, batch_id_node, count_nodes = inputs
 
     # ACSF representation.
     rep_rad = wACSFRad(**w_acsf_rad_kwargs)([node_input, xyz_input, edge_index_input])
@@ -40,6 +42,17 @@ def model_disjoint_weighted(
         out = PoolingNodes(**node_pooling_args)([count_nodes, n, batch_id_node])
         if use_output_mlp:
             out = MLP(**output_mlp)(out)
+        if predict_dipole:
+            pc = Dense(units=1)(n)
+            tc = PoolingNodes(pooling_method="sum")([count_nodes, pc, batch_id_node])
+            if tot_charge is not None:
+                pc_correct = CorrectPartialCharges()([tc, tot_charge, count_nodes, batch_id_node])
+                pc = Add()([pc, pc_correct])
+            p_dip = Multiply()([pc, xyz_input])
+            dip = PoolingNodes(pooling_method="sum")([count_nodes, p_dip, batch_id_node])
+            out = [out, dip]
+            if tot_charge is None:
+                out = out + [tc]
     elif output_embedding == 'node':
         out = n
         if use_output_mlp:
@@ -60,10 +73,11 @@ def model_disjoint_behler(
         mlp_kwargs: dict = None,
         output_embedding: str = None,
         use_output_mlp: bool = None,
-        output_mlp: dict = None
+        output_mlp: dict = None,
+        predict_dipole: bool = None
 ):
     # Make input
-    node_input, xyz_input, edge_index_input, angle_index_input, batch_id_node, count_nodes = inputs
+    node_input, xyz_input, edge_index_input, angle_index_input, tot_charge, batch_id_node, count_nodes = inputs
 
     # ACSF representation.
     rep_g2 = ACSFG2(**ACSFG2.make_param_table(**g2_kwargs))([node_input, xyz_input, edge_index_input])
@@ -84,6 +98,17 @@ def model_disjoint_behler(
         out = PoolingNodes(**node_pooling_args)([count_nodes, n, batch_id_node])
         if use_output_mlp:
             out = MLP(**output_mlp)(out)
+        if predict_dipole:
+            pc = Dense(units=1)(n)
+            tc = PoolingNodes(pooling_method="sum")([count_nodes, pc, batch_id_node])
+            if tot_charge is not None:
+                pc_correct = CorrectPartialCharges()([tc, tot_charge, count_nodes, batch_id_node])
+                pc = Add()([pc, pc_correct])
+            p_dip = Multiply()([pc, xyz_input])
+            dip = PoolingNodes(pooling_method="sum")([count_nodes, p_dip, batch_id_node])
+            out = [out, dip]
+            if tot_charge is None:
+                out = out + [tc]
     elif output_embedding == 'node':
         out = n
         if use_output_mlp:
@@ -100,10 +125,11 @@ def model_disjoint_atom_wise(
         mlp_kwargs: dict = None,
         output_embedding: str = None,
         use_output_mlp: bool = None,
-        output_mlp: dict = None
+        output_mlp: dict = None,
+        predict_dipole: bool = None
 ):
     # Make input
-    node_input, rep_input, batch_id_node, count_nodes = inputs
+    node_input, rep_input, tot_charge, batch_id_node, count_nodes = inputs
 
     # learnable NN.
     n = RelationalMLP(**mlp_kwargs)([rep_input, node_input, batch_id_node, count_nodes])
@@ -113,6 +139,17 @@ def model_disjoint_atom_wise(
         out = PoolingNodes(**node_pooling_args)([count_nodes, n, batch_id_node])
         if use_output_mlp:
             out = MLP(**output_mlp)(out)
+        if predict_dipole:
+            pc = Dense(units=1)(n)
+            tc = PoolingNodes(pooling_method="sum")([count_nodes, pc, batch_id_node])
+            if tot_charge is not None:
+                pc_correct = CorrectPartialCharges()([tc, tot_charge, count_nodes, batch_id_node])
+                pc = Add()([pc, pc_correct])
+            p_dip = Multiply()([pc, xyz_input])
+            dip = PoolingNodes(pooling_method="sum")([count_nodes, p_dip, batch_id_node])
+            out = [out, dip]
+            if tot_charge is None:
+                out = out + [tc]
     elif output_embedding == 'node':
         out = n
         if use_output_mlp: