+++ changelog.md
@@ -14,6 +14,9 @@ causing clashes with built-in functions. We catch defaults to be at least as bac
 * Added ``kgcnn.__safe_scatter_max_min_to_zero__`` for tensorflow and jax backend scattering with default to True.
 * Added simple ragged support for loss and metrics.
 * Added simple ragged support for ``train_force.py``
+* Implemented random equivariant initialize for PAiNN
+* Implemented charge and dipole output for HDNNP2nd
+
 
 v4.0.0